Moved test functions to a testing file
